+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/webcompat/fake/FakeEngineSession.kt
@@ -30,6 +30,19 @@ internal class FakeEngineSession(
         onResult(JSONObject(jsonString))
     }
 
+    override fun sendGleanBrokenSiteReport(
+        details: JSONObject?,
+        description: String?,
+        reason: String,
+        url: String,
+        sendTabSpecificInfo: Boolean,
+        sendBlockedUrls: Boolean,
+        onResult: () -> Unit,
+        onException: (Throwable) -> Unit,
+    ) {
+         onResult()
+    }
+
     override fun getWebCompatInfo(
         onResult: (JSONObject) -> Unit,
         onException: (Throwable) -> Unit,